+++ koin-projects/examples/android-samples/src/test/java/CheckModulesTest.kt
@@ -1,21 +1,19 @@
-import android.app.Activity
+import io.mockk.mockkClass
 import org.junit.Rule
 import org.junit.Test
 import org.junit.experimental.categories.Category
 import org.koin.core.logger.Level
-import org.koin.dsl.module
-import org.koin.sample.androidx.di.allModules
+import org.koin.sample.android.di.allModules
 import org.koin.test.category.CheckModuleTest
 import org.koin.test.check.checkModules
 import org.koin.test.mock.MockProviderRule
-import org.mockito.Mockito
 
 @Category(CheckModuleTest::class)
 class CheckModulesTest {
 
     @get:Rule
     val mockProvider = MockProviderRule.create { clazz ->
-        Mockito.mock(clazz.java)
+        mockkClass(clazz)
     }
 
     @Test