+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/parameter/DefinitionParameters.kt
@@ -18,6 +18,8 @@ package org.koin.core.parameter
 import org.koin.core.error.DefinitionParameterException
 import org.koin.core.error.NoParameterFoundException
 import org.koin.core.parameter.DefinitionParameters.Companion.MAX_PARAMS
+import org.koin.ext.getFullName
+import kotlin.reflect.KClass
 
 /**
  * DefinitionParameters - Parameter holder
@@ -26,16 +28,17 @@ import org.koin.core.parameter.DefinitionParameters.Companion.MAX_PARAMS
  * @author - Arnaud GIULIANI
  */
 @Suppress("UNCHECKED_CAST")
-class DefinitionParameters internal constructor(vararg val values: Any?) {
+open class DefinitionParameters(vararg val values: Any?) {
 
-    private fun <T> elementAt(i: Int): T =
-        if (values.size > i) values[i] as T else throw NoParameterFoundException("Can't get parameter value #$i from $this")
+    open fun <T> elementAt(i: Int, clazz: KClass<*>): T =
+        if (values.size > i) values[i] as T else throw NoParameterFoundException(
+            "Can't get injected parameter #$i from $this for type '${clazz.getFullName()}'")
 
-    operator fun <T> component1(): T = elementAt(0)
-    operator fun <T> component2(): T = elementAt(1)
-    operator fun <T> component3(): T = elementAt(2)
-    operator fun <T> component4(): T = elementAt(3)
-    operator fun <T> component5(): T = elementAt(4)
+    inline operator fun <reified T> component1(): T = elementAt(0, T::class)
+    inline operator fun <reified T> component2(): T = elementAt(1, T::class)
+    inline operator fun <reified T> component3(): T = elementAt(2, T::class)
+    inline operator fun <reified T> component4(): T = elementAt(3, T::class)
+    inline operator fun <reified T> component5(): T = elementAt(4, T::class)
 
     /**
      * get element at given index
@@ -43,7 +46,7 @@ class DefinitionParameters internal constructor(vararg val values: Any?) {
      */
     operator fun <T> get(i: Int) = values[i] as T
 
-    fun <T> set(i: Int,t : T) {
+    fun <T> set(i: Int, t: T) {
         values.toMutableList()[i] = t
     }
 
@@ -71,6 +74,8 @@ class DefinitionParameters internal constructor(vararg val values: Any?) {
     companion object {
         const val MAX_PARAMS = 5
     }
+
+    override fun toString(): String = "DefinitionParameters${values.toList()}"
 }
 
 /**
@@ -80,12 +85,14 @@ class DefinitionParameters internal constructor(vararg val values: Any?) {
  * return ParameterList
  */
 fun parametersOf(vararg parameters: Any?) =
-    if (parameters.size <= MAX_PARAMS) DefinitionParameters(*parameters) else throw DefinitionParameterException("Can't build DefinitionParameters for more than $MAX_PARAMS arguments")
+    if (parameters.size <= MAX_PARAMS) DefinitionParameters(*parameters) else throw DefinitionParameterException(
+        "Can't build DefinitionParameters for more than $MAX_PARAMS arguments")
 
 /**
  *
  */
 fun emptyParametersHolder() = DefinitionParameters()
+
 
 /**
  * Help define a DefinitionParameters